# api/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Ignition API
  version: 0.1.0
  description: >
    HTTP/JSON view of api/proto/ignition/v1 (SandboxService + OperationService).
    Process create/get/list/attach/signal/cancel are on SandboxService.
    Exec bytes travel on ignition-gateway after AttachProcess.
    Implemented behavior: docs/guides/ignition-implementation.md
    Full v1 contract: docs/design/ignition-sandbox-create-api.md
servers:
  - url: https://api.ignition.dev
security:
  - bearerAuth: []
paths:
  /v1/projects/{project}/sandboxes:
    get:
      operationId: listSandboxes
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: pageSize
          in: query
          schema: { type: integer }
        - name: pageToken
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List sandboxes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListSandboxesResponse" }
    post:
      operationId: createSandbox
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSandboxRequest"
      responses:
        "202":
          description: Admission accepted; sandbox is CREATING.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSandboxResponse"
  /v1/projects/{project}/runtimes/default:
    get:
      operationId: getDefaultRuntime
      description: >-
        The system-managed default runtime. Its fields fill any RuntimeSpec
        value a CreateSandbox request leaves unset. Read-only.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Default runtime spec
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuntimeSpec" }
  /v1/projects/{project}/sandboxes/{sandbox}:
    get:
      operationId: getSandbox
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/SandboxId"
      responses:
        "200":
          description: Sandbox resource
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sandbox" }
  /v1/projects/{project}/sandboxes/{sandbox}:terminate:
    post:
      operationId: terminateSandbox
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/SandboxId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Termination accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateSandboxResponse" }
  /v1/projects/{project}/sandboxes/{sandbox}:watch:
    get:
      operationId: watchSandbox
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/SandboxId"
        - name: Last-Event-ID
          in: header
          schema: { type: string }
      responses:
        "200":
          description: SSE stream of Sandbox snapshots
  /v1/projects/{project}/sandboxes/{sandbox}/processes:
    get:
      operationId: listProcesses
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/SandboxId"
        - name: pageSize
          in: query
          schema: { type: integer }
        - name: pageToken
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List processes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListProcessesResponse" }
    post:
      operationId: createProcess
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/SandboxId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcessRequest"
      responses:
        "200":
          description: Process resource (state CREATING or STARTING)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Process"
  /v1/projects/{project}/sandboxes/{sandbox}/processes/{process}:
    get:
      operationId: getProcess
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/SandboxId"
        - $ref: "#/components/parameters/ProcessId"
      responses:
        "200":
          description: Process resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Process"
  /v1/projects/{project}/sandboxes/{sandbox}/processes/{process}:attach:
    post:
      operationId: attachProcess
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/SandboxId"
        - $ref: "#/components/parameters/ProcessId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Exec stream token for ignition-gateway
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachProcessResponse"
  /v1/projects/{project}/sandboxes/{sandbox}/processes/{process}:signal:
    post:
      operationId: signalProcess
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/SandboxId"
        - $ref: "#/components/parameters/ProcessId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [signal]
              properties:
                signal: { type: string }
      responses:
        "200":
          description: Process resource after signal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Process" }
  /v1/projects/{project}/sandboxes/{sandbox}/processes/{process}:cancel:
    post:
      operationId: cancelProcess
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/SandboxId"
        - $ref: "#/components/parameters/ProcessId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Process cancellation accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Process" }
  /v1/projects/{project}/operations:
    get:
      operationId: listOperations
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: List operations
  /v1/projects/{project}/operations/{operation}:
    get:
      operationId: getOperation
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/OperationId"
      responses:
        "200":
          description: Operation resource
  /v1/projects/{project}/operations/{operation}:watch:
    get:
      operationId: watchOperation
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/OperationId"
        - name: Last-Event-ID
          in: header
          schema: { type: string }
      responses:
        "200":
          description: SSE stream of Operation snapshots
  /v1/projects/{project}/operations/{operation}:cancel:
    post:
      operationId: cancelOperation
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/OperationId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Cancel accepted or already terminal
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: RFC 9068 JWT
  parameters:
    ProjectId:
      name: project
      in: path
      required: true
      schema: { type: string }
    SandboxId:
      name: sandbox
      in: path
      required: true
      schema: { type: string }
    OperationId:
      name: operation
      in: path
      required: true
      schema: { type: string }
    ProcessId:
      name: process
      in: path
      required: true
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
  schemas:
    CreateSandboxRequest:
      type: object
      required: [imageId]
      description: >-
        resources, placement, timeouts, and network are optional. Any field
        left unset is filled from the system default runtime
        (GET /v1/projects/{project}/runtimes/default). The resolved RuntimeSpec
        is snapshotted onto the sandbox.
      properties:
        name: { type: string }
        imageId: { type: string }
        command:
          type: array
          items: { type: string }
        workingDirectory: { type: string }
        resources: { $ref: "#/components/schemas/ResourceSpec" }
        placement: { $ref: "#/components/schemas/PlacementSpec" }
        timeouts: { $ref: "#/components/schemas/TimeoutSpec" }
        network: { $ref: "#/components/schemas/NetworkSpec" }
        secretRefs:
          type: array
          items:
            type: object
            required: [secretId, environmentName]
            properties:
              secretId: { type: string }
              version: { type: string }
              environmentName: { type: string }
        labels:
          type: object
          additionalProperties: { type: string }
    CreateSandboxResponse:
      type: object
      required: [sandbox, operation]
      properties:
        sandbox:
          $ref: "#/components/schemas/Sandbox"
        operation:
          $ref: "#/components/schemas/Operation"
    ResourceSpec:
      type: object
      properties:
        cpuMilli: { type: integer, minimum: 1, maximum: 8000 }
        memoryMiB: { type: integer, minimum: 1, maximum: 32768 }
        accelerator:
          type: object
          properties:
            type:
              type: string
              enum: [NONE, NVIDIA_L4]
              description: >-
                Public JSON of proto AcceleratorType (ACCELERATOR_TYPE_ prefix
                stripped). NONE is a CPU-only sandbox.
            count:
              type: integer
              description: NONE requires 0; NVIDIA_L4 requires 1.
    TimeoutSpec:
      type: object
      properties:
        startupSeconds: { type: integer, minimum: 1, maximum: 600 }
        maximumRuntimeSeconds: { type: integer, minimum: 1, maximum: 86400 }
        idleSeconds: { type: integer, minimum: 1, maximum: 3600 }
        terminationGraceSeconds: { type: integer, minimum: 1, maximum: 120 }
    RuntimeSpec:
      type: object
      description: >-
        A sandbox's run configuration. Returned by GetDefaultRuntime; its fields
        fill any value a CreateSandbox request leaves unset.
      properties:
        resources: { $ref: "#/components/schemas/ResourceSpec" }
        placement: { $ref: "#/components/schemas/PlacementSpec" }
        timeouts: { $ref: "#/components/schemas/TimeoutSpec" }
        network: { $ref: "#/components/schemas/NetworkSpec" }
    NetworkSpec:
      type: object
      properties:
        internetAccess:
          type: string
          enum: [DISABLED, ENABLED]
          default: DISABLED
          description: Controls outbound public internet access. It does not enable inbound access or access to private control networks.
    PlacementSpec:
      type: object
      properties:
        region:
          type: string
        computeEnvironment:
          type: string
          enum: [STANDARD, BARE_METAL]
          default: STANDARD
          description: STANDARD uses the managed Ignition sandbox service. BARE_METAL requires a dedicated Compute Engine bare-metal backend.
    Sandbox:
      type: object
      required: [id, projectId, state, imageId, resources]
      properties:
        id: { type: string }
        projectId: { type: string }
        name: { type: string }
        state: { type: string, enum: [CREATING, SCHEDULED, STARTED, READY, TERMINATING, FINISHED, FAILED] }
        stateReason: { type: string }
        imageId: { type: string }
        operationId: { type: string }
        createTime: { type: string, format: date-time }
        readyTime: { type: [string, "null"], format: date-time }
        finishTime: { type: [string, "null"], format: date-time }
        resources: { $ref: "#/components/schemas/ResourceSpec" }
        placement: { $ref: "#/components/schemas/PlacementSpec" }
        timeouts: { $ref: "#/components/schemas/TimeoutSpec" }
        network: { $ref: "#/components/schemas/NetworkSpec" }
        labels: { type: object, additionalProperties: { type: string } }
    Operation:
      type: object
      required: [id, projectId, kind, state, resourceId, createTime]
      properties:
        id: { type: string }
        projectId: { type: string }
        kind: { type: string }
        state: { type: string, enum: [PENDING, RUNNING, SUCCEEDED, FAILED, CANCELLED] }
        resourceId: { type: string }
        createTime: { type: string, format: date-time }
        startTime: { type: [string, "null"], format: date-time }
        endTime: { type: [string, "null"], format: date-time }
        traceId: { type: string }
        progressMessage: { type: string }
    Status:
      type: object
      required: [code, message, requestId, retryable]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
        retryable: { type: boolean }
        retryAfterSeconds: { type: integer, minimum: 0 }
    ListSandboxesResponse:
      type: object
      required: [sandboxes]
      properties:
        sandboxes: { type: array, items: { $ref: "#/components/schemas/Sandbox" } }
        nextPageToken: { type: string }
    ListProcessesResponse:
      type: object
      required: [processes]
      properties:
        processes: { type: array, items: { $ref: "#/components/schemas/Process" } }
    Process:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        sandboxId: { type: string }
        state:
          type: string
          enum: [CREATING, STARTING, RUNNING, CANCELLING, EXITED, FAILED]
        command:
          type: array
          items: { type: string }
        workingDirectory: { type: string }
        pty: { type: boolean }
        exitCode: { type: integer }
        terminatingSignal: { type: string }
    CreateProcessRequest:
      type: object
      required: [command]
      properties:
        command:
          type: array
          items: { type: string }
        workingDirectory: { type: string }
        environment:
          type: object
          additionalProperties: { type: string }
        pty: { type: boolean }
    AttachProcessResponse:
      type: object
      required: [streamToken, gatewayUrl]
      properties:
        streamToken: { type: string }
        gatewayUrl: { type: string }
        expireTime: { type: string, format: date-time }
        streamEpoch: { type: integer, format: int64 }